requirements.txt
fastapi
uvicorn
requests
feedparser
newspaper3k